[FIX] : 🐛 Report CollectionViewCell 페이지별로 분리
